requirements.txt
streamlit>=1.30.0
anthropic>=0.40.0
sqlalchemy>=2.0.0
python-dotenv>=1.0.0
markdown2>=2.4.0
PyMuPDF>=1.23.0
python-docx>=1.0.0
reportlab>=4.0.0
pandas>=2.0.0
